Color scheme choices and shader rework
Rose showed me how to edit the main site colors and so I tinkered with
them until I was happy with the results. Also, significant changes to
shader. Instead of a linear gradient for the halftone, they use more of
a radial gradient 'vignette' effect. The solid colors themselves are
also replaced with simple vertical gradient.